ml/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow==10.4.0
httpx==0.27.2
# Optional heavy dependencies for real generation and scoring
# torch==2.3.1
# torchvision==0.18.1
# diffusers==0.30.0
# transformers==4.43.3
# safetensors==0.4.3
# open-clip-torch==2.24.0
